convert grid to paper layer class
